+++ CHANGELOG.md
@@ -37,6 +37,7 @@
   - You can get closer if you're crouching (50% more distance)
 - Added interaction with the spyglass
   - You can see the birds name and icon when looking at them through the spyglass
+  - Can be turned off in config
 - Hidden config flock settings under a rolldown menu to make it less cluttered
 - Removed bird fade in on spawn
 - Fixed debug scoreboard showing when F3 was enabled